requirements.txt
flask
flask-sqlalchemy
flask-socketio
Werkzeug
Flask-MySQLdb
mysqlclient